+++ CHANGELOG.md
@@ -5,6 +5,10 @@ Enhancements:
 
 * Allow use with Rails 8.0 - [#182](https://github.com/simukappu/activity_notification/pull/182) [#183](https://github.com/simukappu/activity_notification/issues/183)
 
+Bug Fixes:
+
+* Fix mailer jobs failing when notifications are destroyed before the job is performed
+
 ## 2.3.2 / 2024-09-23
 [Full Changelog](http://github.com/simukappu/activity_notification/compare/v2.3.1...v2.3.2)